coredata: Allow setting suproject options from the default_options of the superproject
This only allows the super project to do this, and not for subprojects
to affect each other, as that would create race conditions and just be
really difficult to make work without a two pass interpreter. This
should probably be deprecated in favor of the default_options argument
to subproject() anyway.

Fixes #7573